requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
packaging==24.2
pip==23.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
setuptools==65.5.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
